requirements.txt
asgiref                  3.8.1
cachetools               5.5.0
certifi                  2024.8.30
charset-normalizer       3.4.0
Django                   5.1.3
google-api-core          2.23.0
google-api-python-client 2.154.0
google-auth              2.36.0
google-auth-httplib2     0.2.0
google-auth-oauthlib     1.2.1
googleapis-common-protos 1.66.0
httplib2                 0.22.0
idna                     3.10
oauthlib                 3.2.2
pillow                   11.0.0
pip                      24.0
proto-plus               1.25.0
protobuf                 5.29.1
pyasn1                   0.6.1
pyasn1_modules           0.4.1
pyparsing                3.2.0
requests                 2.32.3
requests-oauthlib        2.0.0
rsa                      4.9
sqlparse                 0.5.2
uritemplate              4.1.1
urllib3                  2.2.3
